+++ src/main/java/io/github/cpjust/testng_annotations/listeners/annotation_transformers/ValueSourceListener.java
@@ -1,4 +1,4 @@
-package io.github.cpjust.testng_annotations.listeners;
+package io.github.cpjust.testng_annotations.listeners.annotation_transformers;
 
 import io.github.cpjust.testng_annotations.annotations.ValueSource;
 import io.github.cpjust.testng_annotations.annotations.NullSource;
@@ -26,11 +26,13 @@ import java.util.function.IntFunction;
 import java.util.function.Supplier;
 
 /**
- * TestNG listener that processes @ValueSource annotations and converts them into data provider parameters.
+ * TestNG listener that processes {@link ValueSource}, {@link NullSource}, {@link EmptySource}, and {@link NullAndEmptySource}
+ * annotations and converts them into data provider parameters.
  */
 @Slf4j
-public class ValueSourceListener implements IAnnotationTransformer {
-    private static final String VALUE_SOURCE_PROVIDER = "valueSourceProvider";
+public class ValueSourceListener extends SourceListenerBase implements IAnnotationTransformer {
+    static final String VALUE_SOURCE_PROVIDER = "valueSourceProvider";
+    static final Map.Entry<Class<?>, String> VALUE_SOURCE_PROVIDER_CLASS_AND_NAME = Map.entry(ValueSourceListener.class, VALUE_SOURCE_PROVIDER);
 
     /**
      * Enum representing the possible value types in @ValueSource.
@@ -49,7 +51,15 @@ public class ValueSourceListener implements IAnnotationTransformer {
     }
 
     /**
-     * Transforms test methods annotated with {@link ValueSource} to use a data provider.
+     * Constructs the listener with the value source data provider.
+     */
+    public ValueSourceListener() {
+        super(List.of(VALUE_SOURCE_PROVIDER_CLASS_AND_NAME));
+    }
+
+    /**
+     * Transforms test methods annotated with {@link ValueSource}, {@link NullSource}, {@link EmptySource}, and
+     * {@link NullAndEmptySource} to use a data provider.
      *
      * @param annotation      The TestNG annotation being transformed.
      * @param testClass       The test class.
@@ -63,17 +73,31 @@ public class ValueSourceListener implements IAnnotationTransformer {
             return; // Nothing to do if there's no test method.
         }
 
-        if (testMethod.isAnnotationPresent(ValueSource.class) ||
-            testMethod.isAnnotationPresent(NullSource.class) ||
-            testMethod.isAnnotationPresent(EmptySource.class) ||
-            testMethod.isAnnotationPresent(NullAndEmptySource.class)) {
-            annotation.setDataProvider(VALUE_SOURCE_PROVIDER);
-            annotation.setDataProviderClass(ValueSourceListener.class);
+        throwIfDataProviderNotAllowed(annotation, testMethod);
+        throwIfTestHasMultipleDataProviders(testMethod);
+
+        if (isValueSourcePresent(testMethod)) {
+            annotation.setDataProvider(VALUE_SOURCE_PROVIDER_CLASS_AND_NAME.getValue());
+            annotation.setDataProviderClass(VALUE_SOURCE_PROVIDER_CLASS_AND_NAME.getKey());
         }
     }
 
     /**
-     * Provides values for test methods annotated with {@link ValueSource} just like a DataProvider would.
+     * Checks if a method is annotated with {@link ValueSource}, {@link NullSource}, {@link EmptySource}, or {@link NullAndEmptySource}.
+     *
+     * @param method The method to check.
+     * @return true if the method is annotated with any of the supported annotations, false otherwise.
+     */
+    public static boolean isValueSourcePresent(@NonNull Method method) {
+        return method.isAnnotationPresent(ValueSource.class) ||
+               method.isAnnotationPresent(NullSource.class) ||
+               method.isAnnotationPresent(EmptySource.class) ||
+               method.isAnnotationPresent(NullAndEmptySource.class);
+    }
+
+    /**
+     * Provides values for test methods annotated with {@link ValueSource}, {@link NullSource}, {@link EmptySource}, and/or
+     * {@link NullAndEmptySource} just like a DataProvider would.
      *
      * @param method The test method.
      * @return An array of parameter values for the test method.