+++ src/test/java/io/github/cpjust/testng_annotations/listeners/annotation_transformers/ValueSourceIT.java
@@ -1,4 +1,4 @@
-package io.github.cpjust.testng_annotations.listeners;
+package io.github.cpjust.testng_annotations.listeners.annotation_transformers;
 
 import io.github.cpjust.testng_annotations.annotations.ValueSource;
 import lombok.extern.slf4j.Slf4j;